Added incomplete script to scrape data from https://apps.bbmpgov.in/covid19/index.html
The script parses the iframe src in this page which is a link to powerbi app.
Then it calls the powerbi url in puppeteer and scrolls all the way down to fetch all the data.
However, when new data is added, old data is removed, therefore need a way to circumvent this issue.